API/requirements.txt
gunicorn
fastapi
uvicorn[standard]
python-dotenv
pymssql
pyodbc 